Made the character manager aware of the number of oranges that have been collected by adding a pointer to the score count for both capy and frog (as well as hybrid) they both increment this on collision

+++ src/CharacterFiles/CharacterManager.h
@@ -29,7 +29,7 @@ public:
     Vector2 initCapyPos;
     Vector2 initFrogPos;
 
-    float numOrangesCollected = 0;
+    int numOrangesCollected = 0;
 
     bool capyAndFrogAreJoined = false;
